Remove portfolio
This means we can also remove angular, since it's basically a static html site
now.

I tried to clean up the relevant stuff, but who knows if I got it all. It'll be
moot soon since I plan to wipe a clean slate to restart with next.js.

Also while I'm here fix various things, not least of which the package.json
metadata.